Bug 1823848 - Part 2: Add SearchSelectorMenuBinding to handle updating the menu items of the search selector menu
